refactor: create io mixin
