+++ docarray/base_document/mixins/__init__.py
@@ -1,4 +1,4 @@
-from docarray.base_document.mixins.proto import ProtoMixin
+from docarray.base_document.mixins.io import IOMixin
 from docarray.base_document.mixins.update import UpdateMixin
 
-__all__ = ['ProtoMixin', 'UpdateMixin']
+__all__ = ['IOMixin', 'UpdateMixin']
